requirements.txt
pandas
numpy
boto3
google-cloud-storage
azure-storage-blob
pymongo
kafka-python
aioredis
redis
neo4j
sqlalchemy
psycopg2-binary
transformers
torch
pytorch-lightning
tensorflow
tensorflow-recommenders
sagemaker
google-cloud-aiplatform
azureml-core
mlflow
wandb
scipy
scikit-learn
streamlit
plotly
networkx
lifetimes
matplotlib
seaborn
pytest
locust
jupyter
kaggle
tqdm
implicit
pyarrow
dask
distributed
